refactor: separate sports and events
