+++ src/app/(screens)/cl-events.tsx
@@ -5,7 +5,6 @@ import { useCallback, useEffect } from 'react'
 import { InteractionManager, View } from 'react-native'
 import { createStyleSheet, useStyles } from 'react-native-unistyles'
 import ClEventsPage from '@/components/Events/cl-events-page'
-import LoadingIndicator from '@/components/Universal/loading-indicator'
 import { useTransparentHeaderPadding } from '@/hooks/useTransparentHeader'
 import { loadCampusLifeEvents, QUERY_KEYS } from '@/utils/events-utils'
 import { pausedToast } from '@/utils/ui-utils'
@@ -13,8 +12,7 @@ import { pausedToast } from '@/utils/ui-utils'
 export default function ClEventsScreen(): React.JSX.Element {
 	const { styles, theme } = useStyles(stylesheet)
 	const headerPadding = useTransparentHeaderPadding()
-	const { tab, openEvent, id } = useLocalSearchParams<{
-		tab?: string
+	const { openEvent, id } = useLocalSearchParams<{
 		openEvent?: string
 		id?: string
 	}>()
@@ -25,19 +23,6 @@ export default function ClEventsScreen(): React.JSX.Element {
 		staleTime: 1000 * 60 * 60, // 60 minutes
 		gcTime: 1000 * 60 * 60 * 24 // 24 hours
 	})
-
-	useEffect(() => {
-		if (tab !== 'sports') {
-			return
-		}
-		router.replace({
-			pathname: '/sports',
-			params: {
-				...(openEvent != null ? { openEvent } : {}),
-				...(id != null ? { id } : {})
-			}
-		})
-	}, [tab, openEvent, id])
 
 	useEffect(() => {
 		if (clEventsResult.isPaused && clEventsResult.data != null) {
@@ -63,13 +48,7 @@ export default function ClEventsScreen(): React.JSX.Element {
 		<View
 			style={[styles.page, { paddingTop: headerPadding + theme.margins.page }]}
 		>
-			{tab === 'sports' ? (
-				<View style={styles.loaderWrapper}>
-					<LoadingIndicator />
-				</View>
-			) : (
-				<ClEventsPage clEventsResult={clEventsResult} />
-			)}
+			<ClEventsPage clEventsResult={clEventsResult} />
 		</View>
 	)
 }
@@ -77,10 +56,5 @@ export default function ClEventsScreen(): React.JSX.Element {
 const stylesheet = createStyleSheet(() => ({
 	page: {
 		flex: 1
-	},
-	loaderWrapper: {
-		flex: 1,
-		justifyContent: 'center',
-		alignItems: 'center'
 	}
 }))
